[quidditch_snitch] Introduce and use `microkernel_fence` op (#81)
Since the introduction of reduction tiling, we've introduced errors in
the kernel code through the introduction of race conditions. We
incorrectly assumed that `barrier` ops serve as synchronization points
for the FPU and the integer core, but this is not the case. This
resulted in scenarios such as the DMA copying an L1 buffer back to DRAM
despite the calculation not having finished or (in the reduction tiling
case) the DMA overriding an L1 buffer that is still in use.

The new `microkernel_fence` op ensures the FPU (and therefore the
microkernel invocations) are in sync with the integer core. The
specialization pass conservatively places the op before
`start_dma_transfer` operations, similar to the barrier.

This will hurt performance compared to the numbers we've previously been
getting, but being fast and wrong is not really useful.