[SpecializeDMACode] Replace `start_dma_transfer` rather than erase (#94)
The previous code simply erased the operation, relying on it only being
used by `wait_for_dma_transfers` that are also getting erased. This very
quickly stops being the case as soon as control flow operations are
involved and is just generally not composable.

The new implementation instead lowers it to a `completed_token` op which
in the compute function should optimize away in later stages.