[SpecializeDMACode] Properly lower `compute_core_index` (#109)
Barriers may be inserted in a loop resulting from the lowering of a
`scf.forall`. The lowering of `compute_core_index` unfortunately returns
`num_compute_cores + 1` which is outside the specified range of the
operation and leads to such loops and their barriers being skipped by
the DMA core. As the pass already assumes non-divergent control flow, we
can fix this by specializing `compute_core_index` to any integer in the
defined range of the op.