[TensorTile] Merge reduction tiling into L1 tiling (#125)
Conceptually, reduction tiling achieves the same goal as L1 tiling: Both
ensure that the output tile is in L1 memory. The main reason why this
used to be an extra step nevertheless is the performance requirement of
tiling reduction prior to any of the parallel dimensions in a matvec to
reduce memory transfers. By adding a loop interchange configuration
(which could in practice also be hard-coded probably), we can replicate
this behavior using only a single tiling config.