[DMA] Split DMA operations into its own dialect (#127)
`quidditch_snitch` has turned into a bit of a god-dialect which even
contains things not specific to snitch. This PR is the first step in
reducing this by splitting everything DMA related out of the dialect.
Future goal is to have lowerings of just DMA operations to lower-level
hardware specific dialects where hardware specific optimizations and
legalizations can occur.